+++ bzfs_main/snapshot_cache.py
@@ -138,6 +138,7 @@ from bzfs_main.utils import (
     DIR_PERMISSIONS,
     LOG_TRACE,
     SortedInterner,
+    sha256_urlsafe_base64,
     stderr_to_str,
 )
 
@@ -180,8 +181,9 @@ class SnapshotCache:
     def last_modified_cache_file(self, remote: Remote, dataset: str, label: SnapshotLabel | None = None) -> str:
         """Returns the path of the cache file that is tracking last snapshot modification."""
         cache_file: str = DATASET_CACHE_FILE_PREFIX if label is None else f"{label.prefix}{label.infix}{label.suffix}"
-        userhost_dir: str = remote.cache_namespace()
-        return os.path.join(self.job.params.log_params.last_modified_cache_dir, userhost_dir, dataset, cache_file)
+        userhost_dir: str = sha256_urlsafe_base64(remote.cache_namespace(), padding=False)
+        dataset_dir: str = sha256_urlsafe_base64(dataset, padding=False)
+        return os.path.join(self.job.params.log_params.last_modified_cache_dir, userhost_dir, dataset_dir, cache_file)
 
     def invalidate_last_modified_cache_dataset(self, dataset: str) -> None:
         """Resets the timestamps of top-level cache files of the given dataset to zero.